Chore: Create a shoutout helper to reduce the amout of code on the shout out service file
