+++ services/slackServices/shoutout.js
@@ -1,7 +1,6 @@
 const { slackModel } = require('../../data/slackModels/slack');
 const Organization = require('../../data/dbModels/organizations');
-const ShoutOut = require('../../data/dbModels/shoutouts');
-const User = require('../../data/dbModels/users');
+const ShoutOutHelper = require('./shoutout.helpers');
 
 exports.sendShoutOut = async reqInfo => {
   try {
@@ -88,36 +87,6 @@ exports.respondToInteractiveMessage = async reqInfo => {
   }
 };
 
-async function getUser(userSlackId, orgId) {
-  try {
-    let user = await User.readBySlackId(userSlackId);
-    let { id } = await Organization.read(orgId);
-    if (!user) {
-      const userData = {
-        org_id: id,
-        slack_mem_id: userSlackId,
-      };
-      user = await User.create(userData);
-      return user;
-    }
-    return user;
-  } catch (err) {
-    console.log(err);
-  }
-}
-
-async function saveToDatabase(dbInfo) {
-  const reciever = await getUser(dbInfo.receiver_id, dbInfo.organization_id);
-  const giver = await getUser(dbInfo.giver_id, dbInfo.organization_id);
-
-  const shoutoutData = {
-    giver_id: giver.id,
-    receiver_id: reciever.id,
-    message: dbInfo.message,
-  };
-  await ShoutOut.create(shoutoutData);
-}
-
 exports.submitDialog = async reqInfo => {
   try {
     const org = await Organization.read(reqInfo.team);
@@ -167,7 +136,7 @@ exports.submitDialog = async reqInfo => {
       receiver_id: reqInfo.recipient,
       message: reqInfo.content,
     };
-    await saveToDatabase(dbInfo);
+    await ShoutOutHelper.saveToDatabase(dbInfo);
   } catch (err) {
     console.log(err);
   }
